requirements.txt
fastapi
streamlit
pytorch-lightning
transformers
tokenizers

elyra>=2.0.1
